+++ redvypr/CHANGELOG.md
@@ -15,9 +15,12 @@ redvypr changelog
 - Metadata changed signal implemented `metadata_changed_signal` in `Redvypr`
 - `RedvyprAddress` can now have the datakey `!`, which means that the datakey must be strictly empty
 - `RedvyprAddress` can compare datetimes `RedvyprAddress("@calibration_date <= dt(2026-01-14T16:15:15)")`
+- `RedvyprAddress` understands quoted strings (does not treat them), see also new test addresses in `test_redvypr_address` 
 - `data_packets.create_datadict` new parameter `random_host="somehostname"` to create a test packet
 - Added maximum size and automatic new file creation to `RedvyprSqliteDb`
 - Added `sensor_calibration_manager` and `sensor_and_calibration_definitions` that allow to pair sensor definitions with calibration definitions
+- Added `__version__` to `__init__.py`
+- Added `test_redvypr_address_benchmark.py`
 
 ### Changed
 - Improved the first page of redvypr, which is now with a timer that starts redvypr after 10 seconds.